requirements.txt
# requirements.txt - UPDATED VERSION
streamlit>=1.28.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
google-generativeai>=0.6.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
rank-bm25>=0.2.2
rapidfuzz>=3.6.1
SQLAlchemy>=2.0.29
pydantic>=2.7.0
streamlit-aggrid>=1.0.5
streamlit-echarts>=0.4.0

# Additional dependencies for better performance
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For OCR support
# pytesseract>=0.3.10
# pillow>=10.0.0
